clean unwanted code
