+++ app/src/main/java/com/lazymohan/zebraprinter/grn/domain/usecase/CreateToReceiptUseCase.kt
@@ -7,7 +7,6 @@ import javax.inject.Inject
 class CreateToReceiptUseCase @Inject constructor(
     private val repo: GrnRepository
 ) {
-    // It can be PO ReceiptRequest or TO payload
     suspend operator fun invoke(body: Any): Result<ReceiptResponse> =
         repo.createReceipt(body)
 }
